Fix include
- Change tempCelsius to Int

+++ app/src/main/java/com/pratyakshkhurana/weatherapp/MainActivity.kt
@@ -125,7 +125,7 @@ class MainActivity : AppCompatActivity(), OnSearchViewHistoryItemClicked {
                         val humidity = it.main.humidity.toString()
                         binding.mainWeatherLayout.humidityPercentage.text = "$humidity%"
 
-                        binding.mainWeatherLayout.tempCelsius.text = it.main.temp.toString()
+                        binding.mainWeatherLayout.tempCelsius.text = it.main.temp.toInt().toString()
 
                         binding.mainWeatherLayout.windPercent.text =
                             it.wind.speed.toString() + " m/s"
